Prevent this warning when building:
The OutputPath property is not set for project 'JsBridge.csproj'.  Please check to make sure that you have specified a valid combination of Configuration and Platform for this project.  Configuration='Release'  Platform='iPhone'.  This error may also appear if some other project is trying to follow a project-to-project reference to this project, this project has been unloaded or is not included in the solution, and the referencing project does not build using the same or an equivalent Configuration or Platform.	C:\Windows\Microsoft.NET\Framework\v4.0.30319\Microsoft.Common.targets	609	5	JsBridge